+++ CHANGES.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 2020-10-06
+
+### Added
+
+- The `match()` function.
+
 ## [0.1.3] - 2020-10-04
 
 ### Added
@@ -32,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The initial public release.
 
+[0.1.4]: https://github.com/weavery/clarity.js/compare/0.1.3...0.1.4
 [0.1.3]: https://github.com/weavery/clarity.js/compare/0.1.2...0.1.3
 [0.1.2]: https://github.com/weavery/clarity.js/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/weavery/clarity.js/compare/0.1.0...0.1.1